+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -3,10 +3,12 @@ import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
 import { activityLogController } from "~/modules/activity-logs/activity-logs.js";
 import { authController } from "~/modules/auth/auth.js";
+import { contributorController } from "~/modules/contributors/contributors.js";
 import { groupController } from "~/modules/groups/groups.js";
 import { permissionController } from "~/modules/permissions/permissions.js";
 import { projectApiKeyController } from "~/modules/project-api-keys/project-api-keys.js";
 import { projectGroupController } from "~/modules/project-groups/project-groups.js";
+import { projectPermissionsController } from "~/modules/project-permissions/project-permissions.js";
 import { projectController } from "~/modules/projects/projects.js";
 import { userController, userService } from "~/modules/users/users.js";
 
@@ -22,7 +24,9 @@ const apiV1 = new BaseServerApplicationApi(
 	...authController.routes,
 	...permissionController.routes,
 	...projectGroupController.routes,
+	...projectPermissionsController.routes,
 	...projectController.routes,
+	...contributorController.routes,
 	...userController.routes,
 	...groupController.routes,
 	...projectApiKeyController.routes,