+++ apps/backend/src/modules/project-groups/libs/types/types.ts
@@ -1,4 +1,6 @@
 export {
 	type ProjectGroupCreateRequestDto,
 	type ProjectGroupCreateResponseDto,
+	type ProjectGroupGetAllRequestDto,
+	type ProjectGroupGetAllResponseDto,
 } from "@git-fit/shared";