+++ apps/backend/src/modules/projects/project.controller.ts
@@ -61,7 +61,12 @@ class ProjectController extends BaseController {
 				),
 			method: "POST",
 			path: ProjectsApiPath.ROOT,
-			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.VIEW_ALL_PROJECTS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 			validation: {
 				body: projectCreateValidationSchema,
 			},
@@ -76,7 +81,7 @@ class ProjectController extends BaseController {
 				),
 			method: "DELETE",
 			path: ProjectsApiPath.$ID,
-			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
+			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_ALL_PROJECTS])],
 		});
 
 		this.addRoute({
@@ -88,7 +93,12 @@ class ProjectController extends BaseController {
 				),
 			method: "GET",
 			path: ProjectsApiPath.ROOT,
-			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.VIEW_ALL_PROJECTS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 		});
 
 		this.addRoute({
@@ -100,7 +110,12 @@ class ProjectController extends BaseController {
 				),
 			method: "GET",
 			path: ProjectsApiPath.$ID,
-			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
+			preHandlers: [
+				checkUserPermissions([
+					PermissionKey.VIEW_ALL_PROJECTS,
+					PermissionKey.MANAGE_ALL_PROJECTS,
+				]),
+			],
 		});
 
 		this.addRoute({
@@ -113,7 +128,7 @@ class ProjectController extends BaseController {
 				),
 			method: "PATCH",
 			path: ProjectsApiPath.$ID,
-			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
+			preHandlers: [checkUserPermissions([PermissionKey.MANAGE_ALL_PROJECTS])],
 			validation: {
 				body: projectPatchValidationSchema,
 			},