+++ apps/frontend/src/libs/constants/navigation-items.constant.ts
@@ -6,13 +6,21 @@ const SIDEBAR_ITEMS: NavigationItem[] = [
 		href: AppRoute.PROJECTS,
 		icon: "project",
 		label: "Projects",
-		pagePermissions: [PermissionKey.VIEW_ALL_PROJECTS],
+		pagePermissions: [
+			PermissionKey.VIEW_ALL_PROJECTS,
+			PermissionKey.MANAGE_ALL_PROJECTS,
+		],
 	},
 	{
 		href: AppRoute.ACCESS_MANAGEMENT,
 		icon: "access",
 		label: "Access Management",
 		pagePermissions: [PermissionKey.MANAGE_USER_ACCESS],
+	},
+	{
+		href: AppRoute.CONTRIBUTORS,
+		icon: "contributors",
+		label: "Contributors",
 	},
 	{
 		href: AppRoute.ANALYTICS,