+++ apps/frontend/src/pages/contributors/libs/helpers/get-contributor-rows.helper.ts
@@ -1,0 +1,14 @@
+import { type ContributorGetAllItemResponseDto } from "~/modules/contributors/contributors.js";
+
+import { type ContributorRow } from "../types/types.js";
+
+const getContributorRows = (
+	contributors: ContributorGetAllItemResponseDto[],
+): ContributorRow[] =>
+	contributors.map((contributor) => ({
+		gitEmails: contributor.gitEmails.map((email) => email.email),
+		name: contributor.name,
+		projects: contributor.projects.map((project) => project.name),
+	}));
+
+export { getContributorRows };