+++ apps/frontend/src/pages/project-access-management/libs/types/user-row.type.ts
@@ -1,0 +1,8 @@
+type UserRow = {
+	createdAt: string;
+	id: number;
+	name: string;
+	projectGroups: string[];
+};
+
+export { type UserRow };