+++ packages/shared/src/modules/project-groups/libs/types/project-group-get-all-request-dto.type.ts
@@ -1,0 +1,5 @@
+type ProjectGroupGetAllRequestDto = {
+	id: number;
+};
+
+export { type ProjectGroupGetAllRequestDto };