+++ packages/shared/src/modules/project-groups/libs/types/project-group-get-all-response-dto.type.ts
@@ -1,0 +1,7 @@
+import { type PaginationResponseDto } from "../../../../libs/types/types.js";
+import { type ProjectGroupGetAllItemResponseDto } from "./project-group-get-all-item-response-dto.type.js";
+
+type ProjectGroupGetAllResponseDto =
+	PaginationResponseDto<ProjectGroupGetAllItemResponseDto>;
+
+export { type ProjectGroupGetAllResponseDto };